requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

asgiref==3.3.1; python_version >= '3.5'
coverage==5.3
django-extensions==3.0.9
django==3.1.3
djangorestframework-simplejwt==4.3.0
djangorestframework==3.12.2
docutils==0.16
psycopg2-binary==2.8.6
pyjwt==1.7.1
pyparsing==2.4.7
pytz==2020.4
sqlparse==0.4.1; python_version >= '3.5'